refactor(operator): Factory 메서드 패턴으로 리팩토링
