+++ src/main/java/calculator/domain/operator/Multiplication.java
@@ -1,0 +1,8 @@
+package calculator.domain.operator;
+
+public class Multiplication implements Operator {
+
+    public int operate(int num1, int num2) {
+        return num1 * num2;
+    }
+}
