+++ src/main/java/calculator/domain/operator/Operator.java
@@ -1,0 +1,6 @@
+package calculator.domain.operator;
+
+public interface Operator {
+
+    public int operate(int num1, int num2);
+}
